Confirm forwards planner generating correct route
Build a simple test to ensure that the constrained planner will properly
find a route out of a problematic deep local minima.  Indicates the
problem is that CBS is not generating the correct set of constraints

Specifically, the problem is that I was using node.t, where node was
a Forward_Constraint_Node for computing the cost.  However, node.t is
not the same thing as the number of nodes in the path, as it is capped
at the maximum length of an out path or constraint time + 1.  Therefore
we need another value to track that.  Annoying