Adjust to new max_time threshold
Changed the definition of max_t in constrained planner to 0 for
empty constraints.  I had to do the same for constrained M* now.  This
changed the results of a number of test cases, which I had to go through
and fix.  Note: max_t is the time at which the environment stops evolving,
and has no more constraints that might need to be avoided via delay.
A constraint at time 0, i.e. the initial configuration, is meaningless,
so we don't have to allow for delay.  The change described above was
needed to get the correct cost computation out of constrained planner
for the sum of costs metric